glitchy/requirements.txt
Flask==2.2.5
requests==2.31.0
waitress==2.1.2
python-dotenv==0.21.1